requirements.txt
Flask==2.2.3
Werkzeug>=2.2.2,<3.0
accelerate==0.28.0
bitsandbytes==0.42.0
peft
bagelML
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
diffusers==0.27.1
transformers==4.35.2